refactor: split db/src/worker.rs into worker/ submodules (#310)
The 1296-line worker.rs grew past the ~800-line file cap as the Worker
struct picked up panic-safety guards, progress tracking, eviction, and a
fourth task variant. Split by responsibility, mirroring the `db/src/books/`
pattern:

  worker.rs              31 lines — module root + public re-exports
  worker/types.rs       242 lines — Task / WorkerConfig / Worker struct
                                    + shared helpers (lock_unpoison,
                                    wall_clock_ms)
  worker/queue.rs       199 lines — post / await_completion + the two
                                    RAII map-cleanup guards
  worker/exec.rs         83 lines — run dispatch loop (resource lock +
                                    scan semaphore + terminal projection)
  worker/handlers.rs     76 lines — execute per-task-kind arms (Scan,
                                    ResolveAuthorPhoto, GenerateThumbs,
                                    Test)
  worker/progress.rs     96 lines — progress_snapshot + retention/eviction
                                    + write_terminal_progress +
                                    report_progress
  worker/tests.rs       654 lines — the existing stress tests, unchanged

Public surface (`omnibus_db::worker::{Task, TaskId, TaskOutcome, Worker,
WorkerConfig}`) is preserved via `pub use` in `worker.rs`, so existing
callers in `server/` and `frontend/` compile unchanged.

All 18 worker tests pass (`cargo test -p omnibus-db worker::`), including
the concurrency, poisoning, and eviction stress tests that are the
acceptance gate for the split.

Closes #310
