+++ CLAUDE.md
@@ -76,7 +76,7 @@ hls.rs              — F2.3 HLS transcode cache: `resolve_audiobook`, `get_part
 thumbs.rs           — F1.2 thumbnail pipeline: decode cover → resize to Sm/Md/Lg (2:3) → atomic per-(book,size) WebP writes under OMNIBUS_THUMBS_DIR; FIFO-by-mtime eviction once total cache exceeds OMNIBUS_THUMBS_CAP_BYTES (default 5 GiB). Backs backend.rs's cover-to-WebP serving path
 indexer.rs          — scan → DB indexing, staleness checks (is_stale + reindex); reindex opts into materialize_sidecars. F2.3 `reindex_audiobooks` uses Phase A.5 group_into_books + Phase B parse_groups + sync_audiobooks (writes book_file_parts rows).
 library_layout.rs   — F0.6 canonical layout: slugify, canonical_path, sidecar_cover_for, allocate_canonical_path (collision suffix)
-worker.rs           — single-process Worker primitive: scan_sem (Scan/ScanAudiobooks) + hls_sem (HlsTranscode) semaphores + per-resource keyed mutex; Task::HlsTranscode dispatches into hls::transcode_book; WorkerConfig::hls_concurrency defaults to max(1, num_cpus/2). F1.11 ResolveAuthorPhoto dispatches into author_photos::resolve.
+worker/             — single-process Worker primitive: scan_sem (Scan/ScanAudiobooks) + hls_sem (HlsTranscode) semaphores + per-resource keyed mutex; Task::HlsTranscode dispatches into hls::transcode_book; WorkerConfig::hls_concurrency defaults to max(1, num_cpus/2). F1.11 ResolveAuthorPhoto dispatches into author_photos::resolve. Split into types (Task/Worker/config), queue (post/await_completion + RAII map-cleanup guards), exec (run dispatch loop), handlers (per-task-kind execute arms), progress (snapshot + retention + report_progress), tests
 author_photos.rs    — F1.11 author profile photo cascade (manual → Open Library → letter); injectable `OpenLibraryConfig` for tests (wiremock). First outbound HTTP in the codebase (reqwest, rustls-tls). Also hosts `fetch_remote_image` for the admin "paste URL" path — gated by a `RemoteImageConfig` SSRF guard (#275): resolves the host, blocks loopback / RFC1918 / link-local (incl. AWS IMDS 169.254.169.254) / multicast / IPv6 ULA / site-local / mapped-v4 before any TCP connect, then pins `reqwest` to the validated addresses so DNS rebinding can't slip in a private IP. The `allow_private_addresses` flag is test-only — production AppState construction always uses the strict default.
 migrations/         — numbered SQL migrations embedded via sqlx::migrate!
 ```
